Lots of work getting the gift cards to work correctly for international.  I believe this is working correctly now.  Decided to go with the exchange rate as read in daily mainly because we had to drop Arab states anyway as it's minimum amount is higher than $10 and it just isn't workable.  In Other countires the differences are so minor and in all cases we give just a fraction more than they can use, and this is what we actually /say/ we are doing on the site so I felt I needed to comply with it.

+++ core/src/main/java/org/mindtrails/service/EmailServiceImpl.java
@@ -204,7 +204,7 @@ public class EmailServiceImpl implements EmailService {
     }
 
     @Override
-    public void sendGiftCard(Participant participant, OrderResponse order, int amount) {
+    public void sendGiftCard(Participant participant, OrderResponse order, double amount) {
         // Prepare the evaluation context
         final Context ctx = new Context();
         Email email = getEmailForType("giftCard");
